tests/submission/PhilipTaylor/canvas/initial.reset.clip.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: initial.reset.clip</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
</head><body class="show_output">

<h1><a href="index.initial.html">initial</a>.<a href="index.initial.reset.html">reset</a>.clip</h1>
<p class="desc">Resetting the canvas state resets the current clip region</p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.initial.reset">initial.reset</a>
</li></ul>
</div>

<p class="output">Actual output:</p>
<canvas class="output" height="50" id="c" width="100"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:</p><p><img alt="" class="output expected" id="expected" src="green-100x50.png">
</p><ul id="d"></ul>
<script>
var t = async_test("Resetting the canvas state resets the current clip region");
_addTest(function(canvas, ctx) {

canvas.width = 100;
ctx.rect(0, 0, 1, 1);
ctx.clip();
canvas.width = 100;
ctx.fillStyle = '#0f0';
ctx.fillRect(0, 0, 100, 50);
_assertPixel(canvas, 20,20, 0,255,0,255, "20,20", "0,255,0,255");


});
</script>

<div id="log"></div>
</body></html>
